Correct assumption about the case of heroes2.agg
The original code is using the presence of `HEROES2.AGG` to determine if the
`Complete Installation` dialog should be shown.  That check expected that file
to have a completely uppercase name, even though fheroes2 has no such
requirement.

This new code allows that file to be provided in any case, and reduces the
duplication of the path to the heroes2 resources directory.